implement light sleep power saving with mcp disabled and 'dummy sleep', create screen rc with dev env ready, suport wsl builds

+++ edmund/src/ui/bootscreen.h
@@ -9,7 +9,7 @@ namespace Edmund {
     class BootScreen : public IScreen
     {
     private:
-      int ellapsed_ms = 0, last_tick = 0, current = 0;
+      unsigned long ellapsed_ms = 0, last_tick = 0, current = 0;
     public:
       BootScreen() : IScreen() { }
